pyproject.toml
[tool.poetry]
name = "faceshapex"
version = "0.1.0"
description = ""
authors = ["devin-ai-integration[bot] <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
gradio = "^5.0.2"
tensorflow = "<2.17"
opencv-python-headless = "<4.10"
numpy = "<1.26"
pillow = "^10.4.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
